+++ internal/buildkit/executor/executor.go
@@ -21,7 +21,7 @@ func (l LifecycleExecutor) Execute(ctx context.Context, opts build.LifecycleOpti
 		return err
 	}
 
-	client, err := client.New(ctx, "")
+	client, err := client.New(ctx, opts.DockerHost)
 	if err != nil {
 		return err
 	}
