--- internal/buildkit/v1/image/README.md
@@ -1,6 +1,0 @@
-# PACKAGE `image`
-
-The **Image** packge is responsible to convert any [`stage`](../../stage/README.md) into an Image.
-If the _Stage_ is MultiArchitecture then an Image is built with a single Appropriate Platform.
-
-The Image is strored in memory only, inorder to export the Image to the available formats take a look at [`export`](../../exports/README.md) package.
